+++ src/buffers/TimeshiftBuffer.h
@@ -40,7 +40,7 @@ namespace timeshift {
 
     virtual bool Open(const std::string inputUrl) override;
     virtual void Close() override;
-    virtual int Read(byte *buffer, size_t length) override;
+    virtual ssize_t Read(byte *buffer, size_t length) override;
     virtual int64_t Seek(int64_t position, int whence) override;
 
     virtual bool CanPauseStream() const override
